+++ hutool-core/src/main/java/cn/hutool/core/convert/impl/GenericEnumConverter.java
@@ -8,7 +8,9 @@ import cn.hutool.core.convert.AbstractConverter;
  * @param <E> 枚举类类型
  * @author Looly
  * @since 4.0.2
+ * @deprecated 请使用{@link EnumConverter}
  */
+@Deprecated
 public class GenericEnumConverter<E extends Enum<E>> extends AbstractConverter<E> {
 	private static final long serialVersionUID = 1L;
 
@@ -25,7 +27,13 @@ public class GenericEnumConverter<E extends Enum<E>> extends AbstractConverter<E
 
 	@Override
 	protected E convertInternal(Object value) {
-		return Enum.valueOf(enumClass, convertToStr(value));
+		//noinspection unchecked
+		E enumValue = (E) EnumConverter.tryConvertEnum(value, this.enumClass);
+		if(null == enumValue && false == value instanceof String){
+			// 最后尝试valueOf转换
+			enumValue = Enum.valueOf(this.enumClass, convertToStr(value));
+		}
+		return enumValue;
 	}
 
 	@Override