+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 * 【http  】     支持patch方法（issue#666@Github）
 * 【crypto】     BCUtil支持更加灵活的密钥类型，增加writePemObject方法
 * 【core  】     增加ServiceLoaderUtil
+* 【core  】     增加EnumUtil.getEnumAt方法
+* 【core  】     增强EnumConvert判断能力（issue#I17082@Gitee）
 
 ### Bug修复